requirements.txt
Django>=4.2,<5.0
mysql-connector-python>=8.0
python-chess>=1.999
dj-database-url>=2.2
psycopg2-binary>=2.9
whitenoise>=6.7
